chore(client): rename gql folder to graphql
